Extract NewCommitsToggle and WhitelistedTeams components
